Re-implement the scale-down workflow using the Graph Framework
Added script to drain and remove nodes from the kubernetes cluster
so that they do not show up as `NotReady` in `kubectl get nodes`
after `kubeadm reset`.

Also, removed redundant inputs from the blueprint. For instance,
the `delta` parameter should be enough when scaling-down the
cluster, instead of additionally supplying the nodes' names.