Implement magic mist_machine* inputs
The long list of inputs has been substituted by a single input (a
python dict), which holds all the required information in order to
either provision a new machine or use an existing one. The structure
of the dict is based on the `cloudify.datatypes.mist.MachineParams`
of the cloudify-mist-plugin.

The mist_machine input is a magic input, which is rendered by the
mist.io UI as a single form. Any blueprint can now simply specify
the `mist_machine` input without the need for users to be familiar
with the underlying key-values or data structures.

The `mist_machine_master` and `mist_machine_worker` inputs are used
to represent the kubernetes master and kubernetes worker nodes created
by the install workflow.

Additionally, the blueprint (as a whole) has been improved, its node
types and workflows simplified. Redundant node templates with no use
have been also removed.